offers: Move OfferSQLEntity class to `pcapi.core.offers.models`
