+++ src/pcapi/models/venue_sql_entity.py
@@ -6,6 +6,7 @@ from sqlalchemy.event import listens_for
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import relationship
 
+from pcapi.core.offers.models import OfferSQLEntity
 from pcapi.domain.ts_vector import create_ts_vector_and_table_args
 from pcapi.domain.postal_code.postal_code import PostalCode
 from pcapi.models.venue_type import VenueType
@@ -14,7 +15,6 @@ from pcapi.models.db import Model
 from pcapi.models.has_address_mixin import HasAddressMixin
 from pcapi.models.has_thumb_mixin import HasThumbMixin
 from pcapi.models.needs_validation_mixin import NeedsValidationMixin
-from pcapi.models.offer_sql_entity import OfferSQLEntity
 from pcapi.models.pc_object import PcObject
 from pcapi.models.providable_mixin import ProvidableMixin
 from pcapi.models.versioned_mixin import VersionedMixin
