+++ src/pcapi/sandboxes/scripts/getters/pro_09_mediation.py
@@ -1,5 +1,5 @@
+from pcapi.core.offers.models import OfferSQLEntity
 from pcapi.models.mediation_sql_entity import MediationSQLEntity
-from pcapi.models.offer_sql_entity import OfferSQLEntity
 from pcapi.models.user_sql_entity import UserSQLEntity
 from pcapi.models import VenueSQLEntity
 from pcapi.repository.user_queries import filter_users_with_at_least_one_validated_offerer_validated_user_offerer
