+++ src/pcapi/sandboxes/scripts/getters/pro_10_desk.py
@@ -1,5 +1,5 @@
 from pcapi.core.bookings.models import Booking
-from pcapi.models.offer_sql_entity import OfferSQLEntity
+from pcapi.core.offers.models import OfferSQLEntity
 from pcapi.models.stock_sql_entity import StockSQLEntity
 from pcapi.models.user_sql_entity import UserSQLEntity
 from pcapi.models import VenueSQLEntity
