+++ src/pcapi/sandboxes/scripts/getters/pro_13_reimbursements.py
@@ -1,5 +1,5 @@
 from pcapi.core.bookings.models import Booking
-from pcapi.models.offer_sql_entity import OfferSQLEntity
+from pcapi.core.offers.models import OfferSQLEntity
 from pcapi.models.offerer import Offerer
 from pcapi.models.payment import Payment
 from pcapi.models.stock_sql_entity import StockSQLEntity
